Cleaning up code for pull request
