+++ src/analog_temp.c
@@ -1,27 +1,25 @@
 /*
 
 	FILENAME: 			analog_temp.c
-	DEPENDENCIES:		spi
+	DEPENDENCIES:		spi, pgmspace, delay, stdint
 
 	DESCRIPTION:		Functions for measuring thermistor temperature
 	AUTHORS:			Dylan Vogel, James Sheridan
-	DATE MODIFIED:	    2017-12-28
+	DATE MODIFIED:	    2017-12-29
+
 	NOTE:               TODO: Implement more efficient channel reading algorithm,
                         instead of having to waste SPI frames from waiting for
                         next message
-                        TODO: Figure out why the static program memory functionality
-                        for the lookup table doesn't work
 
 	REVISION HISTORY:
 
+        2017-12-29      DV: Implemented lookup table with flash memory
         2017-12-28      DV: Created file, implemented lookup table, resistance
                         calculation.
 
 */
 
 #include "analog_temp.h"
-#include <uart/log.h>
-#include <util/delay.h>
 
 void init_thermistor(void){
     // Initialize COM_RST pin on 32M1 as ADC CS
@@ -75,40 +73,14 @@ double convert_thermistor_reading(uint16_t adc_reading){
 
 
 double resistance_to_temp(double res){
+    // Convert the measured thermistor (NXRT15XH103FA1B040) resistance to temperature
+    // Lookup table from manufacturer datasheet (pg 31)
+
     int i;
     double diff, slope, temp;
 
-    // Resintances are stored in kilo-ohms
-/*
-    static const float resistances[34] = {
-        195.652,    148.171,    113.347,    87.559, 68.237,
-        53.650,     42.506,     33.892,     27.219, 22.021,
-        17.926,     14.674,     12.081,     10.000, 8.315,
-        6.948,      5.834,      4.917,      4.161,  3.535,
-        3.014,      2.586,      2.228,      1.925,  1.669,
-        1.452,      1.268,      1.110,      0.974,  0.858,
-        0.758,      0.672,      0.596,      0.531
-    };
-    static const int temps[34] = {
-        -40,    -35,    -30,    -25,    -20,
-        -15,    -10,    -5,     0,      05,
-        10,     15,     20,     25,     30,
-        35,     40,     45,     50,     55,
-        60,     65,     70,     75,     80,
-        85,     90,     95,     100,    105,
-        110,    115,    120,    125
-    };
-
-
-    for (i = 0; i < 34; i++){
-        if ((res - resistances[i]) >= 0){
-            diff = res - resistances[i-1];  //should be negative
-            slope = (temps[i] - temps[i-1]) / (resistances[i] - resistances[i-1]);
-            temp = temps[i-1] + (diff * slope);
-            break;
-        }
-    }*/
-
+    // Resistances are stored in kilo-ohms
+    // PROGMEM instructs the compiler to store these values in flash memory
     static const float resistances[34] PROGMEM = {
         195.652,    148.171,    113.347,    87.559, 68.237,
         53.650,     42.506,     33.892,     27.219, 22.021,
@@ -127,7 +99,6 @@ double resistance_to_temp(double res){
         85,     90,     95,     100,    105,
         110,    115,    120,    125
     };
-
 
     for (i = 0; i < 34; i++){
         if ((res - pgm_read_float(&resistances[i])) >= 0){
