+++ src/main.h
@@ -25,8 +25,6 @@ void poll_int(void);
 
 Queue* cmd_queue;
 
-void thermistor_testing(void);
-
 //void pay_rx_callback(uint8_t* data, uint8_t len);
 //void tx_callback(uint8_t*, uint8_t*);
 //void run_cmd(Data cmd);
